+++ yoti-sdk/src/main/java/com/yoti/mobile/android/sdk/exceptions/YotiSDKAppNotInstalledException.java
@@ -1,0 +1,13 @@
+package com.yoti.mobile.android.sdk.exceptions;
+
+public class YotiSDKAppNotInstalledException extends YotiSDKException {
+
+    AppType notInstalledApp;
+
+    public YotiSDKAppNotInstalledException(AppType notInstalledApp, String message) {
+        super(message);
+        this.notInstalledApp = notInstalledApp;
+    }
+}
+
+